+++ src/common/utility/vga2ansi.cpp
@@ -35,7 +35,7 @@ static void CPrint(const char* in)
 	DWORD bytes_written;
 	if (!StdOut)
 		return;
-	WriteFile(StdOut, in, strlen(in), &bytes_written, NULL);
+	WriteFile(StdOut, in, static_cast<int>(strlen(in)), &bytes_written, NULL);
 }
 #else
 static void CPrint(const char* in)
